#3505
This PR implements the mentioned improvement in the issue. Carefully
only changed the int type for counter & iterator variable declarations,
others were kept as is. Also, kept the size_t types as is (though I
think some could also be changed to Py_ssize_t, depending on
preference).


<!-- ^ Please provide a concise and informative title. -->
<!-- ^ Don't put issue numbers in the title, do this in the PR
description below. -->
<!-- ^ For example, instead of "Fixes #12345" use "Introduce new method
to calculate 1 + 2". -->
<!-- v Describe your changes below in detail. -->
<!-- v Why is this change required? What problem does it solve? -->
<!-- v If this PR resolves an open issue, please link to it here. For
example, "Fixes #12345". -->



### :memo: Checklist

<!-- Put an `x` in all the boxes that apply. -->

- [x] The title is concise and informative.
- [x] The description explains in detail what this PR is about.
- [x] I have linked a relevant issue or discussion.
- [ ] I have created tests covering the changes.
- [ ] I have updated the documentation and checked the documentation
preview.

### :hourglass: Dependencies

<!-- List all open PRs that this PR logically depends on. For example,
-->
<!-- - #12345: short description why this is a dependency -->
<!-- - #34567: ... -->

cc @tscrim
    
URL: https://github.com/sagemath/sage/pull/39965
Reported by: Henry Wu
Reviewer(s): Travis Scrimshaw

+++ src/sage/matrix/matrix2.pyx
@@ -10608,7 +10608,7 @@ cdef class Matrix(Matrix1):
             sage: a.density()
             0
         """
-        cdef int x, y, k
+        cdef Py_ssize_t x, y, k
         k = 0
         nr = self.nrows()
         nc = self.ncols()
